% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{.shared-parameters}
\alias{.shared-parameters}
\title{Parameters used in multiple functions}
\arguments{
\item{id}{The unique identifier of the module. This value must match between
the ui and the server.}
}
\description{
These are parameters that are likely to show up in multiple functions, so I
gather their definitions in one easy-to-find place.
}
\keyword{internal}
