% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/social_contacts.R
\name{social_contacts_ui}
\alias{social_contacts_ui}
\alias{social_contacts_server}
\title{Collect social media contact information}
\usage{
social_contacts_ui(id, twitter = TRUE, linkedin = TRUE)

social_contacts_server(id)
}
\arguments{
\item{id}{The unique identifier of the module. This value must match between
the ui and the server.}

\item{twitter, linkedin}{Logical; whether to include the given site in the ui.
At least one of these values must be \code{TRUE}.}
}
\value{
A shiny module ui, or the normalized contacts from one.
}
\description{
A shiny module to collect and normalize a user's social media contact
information.
}
